gpu_price_oracle/requirements.txt
aiohttp>=3.9.0
web3>=6.15.0
python-dotenv>=1.0.0
